refactor: useWorkspaceFsPath
