+++ src/hover-provider.ts
@@ -2,7 +2,7 @@ import type * as vscode from 'vscode'
 
 import type { DecorationV3 } from './decoration-v3'
 import type { DecorationV4 } from './decoration-v4'
-import { enableHoverProvider } from './utils'
+import { enableHoverProvider } from './state'
 
 export class GeneratedCSSHoverProvider implements vscode.HoverProvider {
   constructor(public decoration: DecorationV3 | DecorationV4) {}
